refactor: move hasNonEmptyDeliveredPayloads to blockUtils
- Remove separate blockDebug.ts file
- Move the hasNonEmptyDeliveredPayloads function to blockUtils.ts
- Update import references across all files
