+++ frontend/src/components/beacon/block_production/common/blockUtils.ts
@@ -1,7 +1,42 @@
 /**
  * Helper functions for block-related logic
  */
-import { hasNonEmptyDeliveredPayloads } from './blockDebug';
+
+/**
+ * Checks if the block data has non-empty delivered payloads by looking in multiple places.
+ * Returns true if deliveredPayloads exists and is not empty.
+ */
+export function hasNonEmptyDeliveredPayloads(block: any, parentData?: any): boolean {
+  if (!block && !parentData) {
+    return false;
+  }
+
+  try {
+    // Check in block first
+    if (block && 'deliveredPayloads' in block) {
+      if (typeof block.deliveredPayloads === 'object') {
+        const keys = Object.keys(block.deliveredPayloads);
+        if (keys.length > 0) {
+          return true;
+        }
+      }
+    }
+    
+    // Then check in parent data (slotData)
+    if (parentData && 'deliveredPayloads' in parentData) {
+      if (typeof parentData.deliveredPayloads === 'object') {
+        const keys = Object.keys(parentData.deliveredPayloads);
+        if (keys.length > 0) {
+          return true;
+        }
+      }
+    }
+    
+    return false;
+  } catch (error) {
+    return false;
+  }
+}
 
 /**
  * Determines if a block was locally built by the proposer (not from an external builder).
@@ -19,7 +54,7 @@ export function isBlockLocallyBuilt(block: any): boolean {
     return false;
   }
   
-  // Use special function for a thorough check
+  // Use the helper function for a thorough check
   const hasPayloads = hasNonEmptyDeliveredPayloads(block);
   if (hasPayloads) {
     return false;
